requirements.txt
Flask==2.1.3
graphviz==0.20
gunicorn==20.1.0
Jinja2==3.1.2
librosa==0.9.1
livelossplot==0.5.5
llvmlite==0.38.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
pydot==1.4.2
pydotplus==2.0.2
rsa==4.8
scikit-learn==1.1.1
tensorflow-cpu==2.9.1
Werkzeug==2.1.2